FILE: renames MM local-storage.service.ts to local-state.service.ts
