+++ libs/menu-matriarch/data-access/src/lib/router.service.ts
@@ -10,7 +10,7 @@ import { BehaviorSubject, Observable } from 'rxjs';
 import { map, filter, tap, distinctUntilChanged } from 'rxjs/operators';
 
 import { PlannerView, Route } from '@atocha/menu-matriarch/util';
-import { LocalStorageService } from './internal/local-storage.service';
+import { LocalStorageService } from './internal/local-state.service';
 
 @Injectable({
   providedIn: 'root',
